bloomfilter: Move most Data.BloomFilter modules under Data.BloomFilter.Classic
in preparation for adding blocked bloom filters.

Though we'll do a bit of module consolodation first.
