+++ bloomfilter/src/Data/BloomFilter/Classic/Mutable.hs
@@ -12,9 +12,9 @@
 --   the element is present.
 --
 -- This module provides low-level control.  For an easier to use
--- interface, see the "Data.BloomFilter.Easy" module.
+-- interface, see the "Data.BloomFilter.Classic.Easy" module.
 
-module Data.BloomFilter.Mutable (
+module Data.BloomFilter.Classic.Mutable (
     -- * Overview
     -- $overview
 
@@ -48,8 +48,8 @@ import           Control.Monad.ST (ST)
 import           Data.Kind (Type)
 import           Data.Primitive.ByteArray (MutableByteArray)
 
-import qualified Data.BloomFilter.BitVec64 as V
-import           Data.BloomFilter.Calc (BloomSize (..))
+import qualified Data.BloomFilter.Classic.BitVec64 as V
+import           Data.BloomFilter.Classic.Calc (BloomSize (..))
 import           Data.BloomFilter.Hash (CheapHashes, Hash, Hashable, evalHashes,
                      makeHashes)
 
@@ -143,8 +143,8 @@ size MBloom { numBits, numHashes } =
 --
 -- By choosing these parameters with care, it is possible to tune for
 -- a particular false positive rate.
--- The 'Data.BloomFilter.Easy.suggestSizing' function in
--- the "Data.BloomFilter.Easy" module calculates useful estimates for
+-- The 'Data.BloomFilter.Classic.Easy.suggestSizing' function in
+-- the "Data.BloomFilter.Classic.Easy" module calculates useful estimates for
 -- these parameters.
 
 -- $ease
